fix: exception when using RichText and preview
Widget DraftailRichTextArea will add a default content event if empty
(eg. `<p data-block-key="e47l4"></p>`). When we display the preview,
wagtail try to save the form and a locale seems not empty due to
DraftailRichTextArea value. Exceptions can be raised when some fields
must be unique and stay empty for eg.